backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
mediapipe>=0.10.30
opencv-python>=4.9.0
numpy>=1.26.0
python-multipart>=0.0.6
pillow>=10.2.0
